+++ carpetmodSrc/carpet/logging/logHelpers/ExplosionLogHelper.java
@@ -6,99 +6,63 @@ import net.minecraft.entity.Entity;
 import net.minecraft.util.math.Vec3d;
 import net.minecraft.util.text.ITextComponent;
 
+import java.util.ArrayList;
+import java.util.List;
+
 public class ExplosionLogHelper {
 
     // CARPET-SYLKOS
     // Some code yeeted from lntricarpet and gnembon 1.16+ fabric carpet
 
-    public static Vec3d previousPosition = null;
-    public static long startTime = 0;
-    public static boolean tickHasCompact = false;
-    // TODO: 2023/5/3,0003 To log the blocks with dropping rates
-    private static boolean affectBlocks = false; // will be used later when I add in the full explosion logger
+    // carpet12RNY
+    // Modified by Naftalluvia
+
     private static long lastGametime = 0;
     private static long explosionCountInCurrentGT = 0;
-    private static long explosionCountInCurrentPos = 0;
+    private static boolean newTick = false;
     public final Vec3d pos;
     public final Entity entity;
+    private final float power;
+    private final boolean createFire;
+    // TODO: 2023/5/3,0003 To log the blocks with dropping rates
+    private boolean affectBlocks = false;
 
     public ExplosionLogHelper(Entity entity, double x, double y, double z, float power, boolean createFire) { // blocks removed
         this.entity = entity;
         this.pos = new Vec3d(x, y, z);
+        this.power = power;
+        this.createFire = createFire;
     }
 
-    public static void logLastExplosion() {
-        if (LoggerRegistry.__explosions) {
-            if (tickHasCompact) {
-                tickHasCompact = false;
-                LoggerRegistry.getLogger("explosions").log((option) -> {
-                    ITextComponent[] msg = null;
-                    if ("compact".equals(option)) {
-                        if (previousPosition != null) {
-                            msg = new ITextComponent[]{Messenger.m(null,
-                                    "d #" + (explosionCountInCurrentGT),
-                                    "gb ->",
-                                    "d " + explosionCountInCurrentPos + "x ",
-                                    Messenger.dblt("l", previousPosition.x, previousPosition.y, previousPosition.z),
-                                    (affectBlocks) ? "m (affects blocks)" : "m  (doesn't affect blocks)",
-                                    "g (", "d " + (System.currentTimeMillis() - startTime), "g ms)"
-                            )};
-                        }
-                        startTime = 0;
-                    }
-                    return msg;
-                });
-            }
-        }
+    public void setAffectBlocks(boolean affectBlocks) {
+        this.affectBlocks = affectBlocks;
     }
 
     public void onExplosionDone(long gametime) {
+        newTick = false;
         if (lastGametime != gametime) {
-            explosionCountInCurrentGT = 1;
-            explosionCountInCurrentPos = 0;
-            previousPosition = pos;
+            explosionCountInCurrentGT = 0;
             lastGametime = gametime;
-            startTime = System.currentTimeMillis();
-            LoggerRegistry.getLogger("explosions").log((option) -> {
-                return new ITextComponent[]{Messenger.m(null, "wb tick : ", "d " + gametime)};
-            });
+            newTick = true;
         }
-
+        ++explosionCountInCurrentGT;
         LoggerRegistry.getLogger("explosions").log((option) -> {
-            ITextComponent[] msg = null;
+            List<ITextComponent> messages = new ArrayList<>();
+            if (newTick) {
+                messages.add(Messenger.m(null, "wb tick : ", "d " + gametime));
+            }
             switch (option) {
                 case "brief":
-                    msg = new ITextComponent[]{Messenger.m(null,
+                case "full":
+                    messages.add(Messenger.m(null,
                             "d #" + explosionCountInCurrentGT,
                             "gb ->",
                             Messenger.dblt("l", pos.x, pos.y, pos.z),
-                            (affectBlocks) ? "m (affects blocks)" : "m  (doesn't affect blocks)"
-                    )};
-                    explosionCountInCurrentGT++;
-                    break;
-
-                // temporarily removed "full" because its not really needed for my use case. may implement later - Sylkos
-
-                case "compact":
-                    tickHasCompact = true;
-                    if (previousPosition != null && !pos.equals(previousPosition)) {
-                        msg = new ITextComponent[]{Messenger.m(null,
-                                "d #" + explosionCountInCurrentGT,
-                                "gb ->",
-                                "d " + explosionCountInCurrentPos + "x ",
-                                Messenger.dblt("l", previousPosition.x, previousPosition.y, previousPosition.z),
-                                (affectBlocks) ? "m (affects blocks)" : "m  (doesn't affect blocks)",
-                                "g (", "d " + (System.currentTimeMillis() - startTime), "g ms)"
-                        )};
-                        explosionCountInCurrentGT += explosionCountInCurrentPos;
-                        explosionCountInCurrentPos = 0;
-                        previousPosition = pos;
-                        startTime = System.currentTimeMillis();
-                    }
-                    explosionCountInCurrentPos++;
+                            (affectBlocks ? "m (affects blocks)" : "m  (doesn't affect blocks)")
+                    ));
                     break;
             }
-            return msg;
+            return messages.toArray(new ITextComponent[0]);
         });
     }
 }