+++ ChangeLog.md
@@ -7,3 +7,5 @@ Carpet 1.12 with [RNY](https://github.com/Rainyaphthyl)'s Addition
 ## 1.1. Miscellaneous
 
 1. The logger uses the tick counter with the same output of the command `time query gametime` when displaying the "tick", instead of the wrong one.
+2. The "affects blocks" info can be displayed correctly.
+3. The option "compact" is removed.